feat: Restore GameDetailV2Component with comparison functionality
- Reverted GameDetailV2Component to original working state with view toggle
- Restored three view modes: Original (separate critical points), Inline (Lichess-style), and Both (side-by-side)
- Fixed move navigation and chessboard position updates
- Maintained interactive move clicking with visual feedback
- Preserved critical points analysis and inline annotations
- Added move-list-v2 component template for Lichess-style rendering
- Created game-detail-legacy backup of original component
- Updated project documentation and session notes

This restores the functional comparison interface that allows evaluation
of UX differences between traditional critical points vs inline annotations.
