+++ .env.example
@@ -1,12 +1,16 @@
-# Chess Analysis Platform - Environment Configuration
+# ChessPulse - Multi-User Chess Analysis Platform Environment Configuration
 # Copy this file to .env and configure as needed
-
-# Target player for analysis (default: AdvaitKumar1213)
-# Change this to analyze games for a different player
-TARGET_PLAYER=AdvaitKumar1213
 
 # API Server Port (default: 3000)
 PORT=3000
+
+# Supabase Configuration (required for authentication)
+SUPABASE_URL=your_supabase_url_here
+SUPABASE_ANON_KEY=your_supabase_anon_key_here
+SUPABASE_SERVICE_ROLE_KEY=your_supabase_service_role_key_here
+
+# Database Configuration
+DATABASE_URL=your_database_url_here
 
 # Database Configuration
 # For local development, SQLite is used automatically
