+++ .kiro/project-status.json
@@ -1,0 +1,42 @@
+{
+  "project": "ChessPulse - Multi-User Chess Analysis Platform",
+  "lastUpdated": "2026-01-04T22:52:20.279+05:30",
+  "status": {
+    "authentication": "completed",
+    "gameDetailV2": "completed - now primary",
+    "alternativeMoves": "completed",
+    "documentation": "updated",
+    "componentCleanup": "completed"
+  },
+  "keyComponents": {
+    "backend": {
+      "framework": "Node.js + Express",
+      "database": "SQLite/PostgreSQL + Supabase",
+      "chessEngine": "Stockfish",
+      "authentication": "Supabase"
+    },
+    "frontend": {
+      "framework": "Angular 17",
+      "styling": "Tailwind CSS",
+      "chessBoard": "@lichess-org/chessground",
+      "gameViewer": "Custom Lichess-style PGN viewer"
+    }
+  },
+  "recentChanges": {
+    "removedTargetPlayer": "Migrated from single-user to multi-user authentication",
+    "gameDetailV2": "Promoted to primary component, legacy backed up",
+    "alternativeMovesPattern": "Implemented pgn-viewer style inline variations",
+    "documentationUpdate": "Updated all references to reflect multi-user platform",
+    "routingCleanup": "Removed duplicate v2 route, simplified navigation"
+  },
+  "technicalDebt": [
+    "Clean up TARGET_PLAYER references in test files",
+    "Remove backup files (.bak, .backup)",
+    "Update legacy documentation"
+  ],
+  "nextPriorities": [
+    "Test V2 component with real data",
+    "Performance optimization for alternatives loading",
+    "User feedback collection on new interface"
+  ]
+}
