+++ .kiro/session-2026-01-04.md
@@ -1,0 +1,74 @@
+# Recent Session Summary - January 4, 2026
+
+## Session Focus: Game Detail V2 Component Enhancement
+
+### Completed Tasks
+
+#### 1. Documentation Updates
+- **Updated README.md**: Changed from single-user to multi-user platform description
+- **Updated .env.example**: Removed TARGET_PLAYER, added Supabase configuration
+- **Updated CLAUDE.md**: Replaced hardcoded player references with multi-user authentication info
+
+#### 2. Game Detail V2 Component Feature Parity
+**File**: `/frontend/src/app/pages/game-detail-v2/game-detail-v2.component.ts`
+
+**Added Features**:
+- **Alternative Moves Analysis**: Inline variations following pgn-viewer pattern
+- **Multi-Variation Analysis**: Expandable sections for critical positions  
+- **Game Statistics**: Player accuracy, centipawn loss, move quality distribution
+- **Phase Analysis**: Opening/middlegame/endgame performance breakdown
+- **Enhanced Data Loading**: Async loading of accuracy, phases, and alternatives data
+
+**Key Implementation Details**:
+- `moveAlternatives: Map<number, AlternativeMove[]>` for storing alternatives per move
+- `hasAlternatives(moveIndex)` and `getAlternatives(moveIndex)` helper methods
+- `loadAlternativesForAllMoves()` for preloading alternatives for mistakes/blunders/inaccuracies
+- Inline `<variation>` elements in move list template
+- CSS styling for variation elements with hover effects and best move highlighting
+
+#### 3. Alternative Moves Display Pattern
+**Requirement**: Follow pgn-viewer option #6 for showing alternative moves
+**Implementation**: 
+- Alternatives shown inline within move list as clickable `<variation>` elements
+- Only displayed for mistakes, inaccuracies, and blunders
+- Each variation shows rank badge, move notation, and evaluation
+- Limited to top 3 alternatives per move to avoid clutter
+
+### Technical Approach
+
+#### CSS Grid Layout
+- Maintained Lichess-style responsive design
+- Added variation styling with proper spacing and visual hierarchy
+- Integrated seamlessly with existing move list layout
+
+#### Data Management
+- Efficient Map-based storage for move alternatives
+- Async loading to avoid blocking main game data
+- Proper error handling for API failures
+
+#### User Experience
+- All alternatives clickable with position preview
+- Hover effects for better interaction feedback
+- Maintains keyboard navigation support
+- Responsive design for mobile/desktop
+
+### Files Modified
+1. `/frontend/src/app/pages/game-detail-v2/game-detail-v2.component.ts` - Main component
+2. `/README.md` - Project documentation
+3. `/.env.example` - Environment configuration
+4. `/CLAUDE.md` - Technical documentation
+
+### Next Steps Identified
+- Test the V2 component with real game data
+- Consider performance optimizations for large numbers of alternatives
+- Evaluate user feedback on inline vs panel-based alternative display
+- Clean up remaining TARGET_PLAYER references in test files
+
+### Key Learnings
+- pgn-viewer uses inline variation elements for alternative moves
+- Map-based storage provides efficient lookup for move alternatives
+- Async loading pattern works well for non-critical data like alternatives
+- CSS Grid layout is flexible enough to accommodate inline variations
+
+---
+*Session completed: January 4, 2026 at 22:52 IST*
