+++ frontend/frontend/src/environments/environment.ts
@@ -1,0 +1,10 @@
+export const environment = {
+  production: true,
+  apiUrl: '/api', // Production API URL
+  // Supabase configuration
+  // Get these from https://app.supabase.com -> Project Settings -> API
+  supabaseUrl: 'https://your-project-id.supabase.co', // Replace with your Supabase project URL
+  supabasePublishableKey: 'your_publishable_key_here' // Replace with your Supabase publishable key
+};
+
+
